2023-08-08/requirements.txt
jupyter
numpy
pandas
scikit-learn
sf-hamilton[visualization]
xgboost